[Refactor] Flatten kernels/convolution/ directory into single convolution.py
Merge common.py, conv1d.py, conv2d.py, conv3d.py into one flat module.
No code logic changes — pure file consolidation.
